Introduce game class
Cleanup uses of goto and encapsulate game logic.
